add rate limit on all handlers, refactor/move to handler folder
