+++ src/components/core/compute/getResults.ts
@@ -1,6 +1,6 @@
 import { P2PCommandResponse } from '../../../@types/index.js'
 import { CORE_LOGGER } from '../../../utils/logging/common.js'
-import { Handler } from '../handler.js'
+import { Handler } from '../handler/handler.js'
 import { ComputeGetResultCommand } from '../../../@types/commands.js'
 import { C2DEngine } from '../../c2d/compute_engines.js'
 import { checkNonce, NonceResponse } from '../utils/nonceHandler.js'
