+++ src/components/core/handler/encryptHandler.ts
@@ -1,18 +1,19 @@
 import { Handler } from './handler.js'
-import { P2PCommandResponse } from '../../@types/OceanNode.js'
-import { EncryptCommand, EncryptFileCommand } from '../../@types/commands.js'
+import { P2PCommandResponse } from '../../../@types/OceanNode.js'
+import { EncryptCommand, EncryptFileCommand } from '../../../@types/commands.js'
 import * as base58 from 'base58-js'
 import { Readable } from 'stream'
-import { encrypt } from '../../utils/crypt.js'
-import { Storage } from '../storage/index.js'
-import { getConfiguration } from '../../utils/index.js'
-import { EncryptMethod } from '../../@types/fileObject.js'
+import { encrypt } from '../../../utils/crypt.js'
+import { Storage } from '../../storage/index.js'
+import { getConfiguration } from '../../../utils/index.js'
+import { EncryptMethod } from '../../../@types/fileObject.js'
 import {
   ValidateParams,
   buildInvalidParametersResponse,
   buildInvalidRequestMessage,
+  buildRateLimitReachedResponse,
   validateCommandParameters
-} from '../httpRoutes/validateCommands.js'
+} from '../../httpRoutes/validateCommands.js'
 
 // for encryption
 export const SUPPORTED_ENCRYPTION_ENCODINGS = ['string', 'base58']
@@ -49,6 +50,9 @@ export class EncryptHandler extends Handler {
   }
 
   async handle(task: EncryptCommand): Promise<P2PCommandResponse> {
+    if (!(await this.checkRateLimit())) {
+      return buildRateLimitReachedResponse()
+    }
     const validation = this.validate(task)
     if (!validation.valid) {
       return buildInvalidParametersResponse(validation)
@@ -105,6 +109,9 @@ export class EncryptFileHandler extends Handler {
   }
 
   async handle(task: EncryptFileCommand): Promise<P2PCommandResponse> {
+    if (!(await this.checkRateLimit())) {
+      return buildRateLimitReachedResponse()
+    }
     const validation = this.validate(task)
     if (!validation.valid) {
       return buildInvalidParametersResponse(validation)
