+++ src/components/core/handler/nonceHandler.ts
@@ -1,13 +1,14 @@
 import { Handler } from './handler.js'
-import { P2PCommandResponse } from '../../@types/OceanNode.js'
-import { NonceCommand } from '../../@types/commands.js'
-import { getNonce } from './utils/nonceHandler.js'
+import { P2PCommandResponse } from '../../../@types/OceanNode.js'
+import { NonceCommand } from '../../../@types/commands.js'
+import { getNonce } from '../utils/nonceHandler.js'
 import {
   ValidateParams,
   buildInvalidParametersResponse,
   buildInvalidRequestMessage,
+  buildRateLimitReachedResponse,
   validateCommandParameters
-} from '../httpRoutes/validateCommands.js'
+} from '../../httpRoutes/validateCommands.js'
 import { isAddress } from 'ethers'
 
 export class NonceHandler extends Handler {
@@ -25,6 +26,9 @@ export class NonceHandler extends Handler {
 
   // eslint-disable-next-line require-await
   async handle(task: NonceCommand): Promise<P2PCommandResponse> {
+    if (!(await this.checkRateLimit())) {
+      return buildRateLimitReachedResponse()
+    }
     const validation = this.validate(task)
     if (!validation.valid) {
       return buildInvalidParametersResponse(validation)
