+++ src/components/httpRoutes/aquarius.ts
@@ -3,8 +3,8 @@ import { streamToString } from '../../utils/util.js'
 import { Readable } from 'stream'
 import { PROTOCOL_COMMANDS } from '../../utils/constants.js'
 import { LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
-import { GetDdoHandler, ValidateDDOHandler } from '../core/ddoHandler.js'
-import { QueryHandler } from '../core/queryHandler.js'
+import { GetDdoHandler, ValidateDDOHandler } from '../core/handler/ddoHandler.js'
+import { QueryHandler } from '../core/handler/queryHandler.js'
 import { HTTP_LOGGER } from '../../utils/logging/common.js'
 import { DDO } from '../../@types/DDO/DDO.js'
