+++ src/test/utils/assets.ts
@@ -24,7 +24,7 @@ import { encrypt } from '../../utils/crypt.js'
 import { AssetUtils } from '../../utils/asset.js'
 
 import { PROTOCOL_COMMANDS, getConfiguration } from '../../utils/index.js'
-import { FeesHandler } from '../../components/core/feesHandler.js'
+import { FeesHandler } from '../../components/core/handler/feesHandler.js'
 import { OceanNode } from '../../OceanNode.js'
 import { ProviderFees } from '../../@types/Fees.js'
